added some tests for DnsServer::new()

+++ src/constants.rs
@@ -1,2 +1,2 @@
-pub const BUFFER_MAX_LEN: u64 = 512;
+pub const BUFFER_MAX_LEN: usize = 512;
 pub const UDP_PORT: u64 = 2053;
